# ping360 scan speed optimizations