--- a/src/link/abstractlink.h
+++ b/src/link/abstractlink.h
@@ -212,9 +212,23 @@ public:
     void write(const char* data, int size)
     {
         if(size > 0) {
-            emit sendData(QByteArray(data, size));
-        };
+            if (_type == LinkType::Serial){
+                writeSync(data, size);
+            } else {
+                emit sendData(QByteArray(data, size));
+            }
+        }
     }
+
+    virtual void writeSync(const char* data, int size) { Q_UNUSED(data) Q_UNUSED(size) }
+
+    /**
+     * @brief Copy operation
+     *
+     * @param other
+     * @return const AbstractLink&
+     */
+    const AbstractLink& operator=(const AbstractLink& other);
 
     Q_PROPERTY(qint64 byteSize READ byteSize NOTIFY byteSizeChanged)
     Q_PROPERTY(LinkConfiguration* configuration READ configuration NOTIFY configurationChanged)