--- a/src/sensor/ping360.cpp
+++ b/src/sensor/ping360.cpp
@@ -146,11 +146,6 @@ void Ping360::handleMessage(const ping_message& msg)
 {
     qCDebug(PING_PROTOCOL_PING360) << "Handling Message:" << msg.message_id();
 
-    // Update frequency for each
-    messageFrequencies[msg.message_id()].setElapsed(_messageElapsedTimer.elapsed());
-    // Since we don't have a huge number of messages and this variable is pretty simple,
-    // we can use a single signal to update someone about the frequency update
-    emit messageFrequencyChanged();
 
     switch (msg.message_id()) {
 
@@ -236,6 +231,12 @@ void Ping360::handleMessage(const ping_message& msg)
         break;
     }
     emit parsedMsgsUpdate();
+
+    // Update frequency for each
+    messageFrequencies[msg.message_id()].setElapsed(_messageElapsedTimer.elapsed());
+    // Since we don't have a huge number of messages and this variable is pretty simple,
+    // we can use a single signal to update someone about the frequency update
+    emit messageFrequencyChanged();
 }
 
 void Ping360::firmwareUpdate(QString fileUrl, bool sendPingGotoBootloader, int baud, bool verify)