--- a/src/sensor/pingsensor.h.orig
+++ b/src/sensor/pingsensor.h.orig
@@ -35,8 +35,7 @@ public:
      */
     Q_INVOKABLE virtual void connectLink(AbstractLinkNamespace::LinkType connType, const QStringList& connString) = 0;
 
-    void connectLink(const LinkConfiguration conConf,
-                     const LinkConfiguration& logConf = LinkConfiguration()) override final;
+    void connectLink(const LinkConfiguration conConf, const LinkConfiguration& logConf = LinkConfiguration()) override final;
 
     /**
      * @brief Get device source ID