This is based on #537.

This is the situation that we are dealing with regarding communication between the ping360 device hardware and the ping-viewer application:

![image](https://user-images.githubusercontent.com/8315108/62658988-34674100-b938-11e9-9c69-87ccf5891f4d.png)

This time period of 18 milliseconds is what I will term the `viewer-side latency`. The answer to the question 'is this viewer-side latency necessary?' is **no**! Ideally, it will be zero.

The sonar response with default settings is a constant 8ms. The maximum theoretical message frequency @2Mbaud in this case is 1/0.008 = 125Hz (viewer-side latency = 0). The maximum theoretical full scan speed is 400 / 125 = 3.2 seconds

In an attempt to reduce the signal->slot chain between ping-viewer rx -> ping-viewer tx, I did two primary things:

- removed use of newMessage signal from parser, and parse the buffer/handle the message synchronously
- added writeSync to write immediately when abstractlink::write is called

I also moved some calls around inside of `handleMessage` in order to request the next data message from the ping360 device as soon as possible.

**master communicating at default 2m range and 2Mbaud scans at 30 Hz**
**this patch communicating at default 2m range and 2Mbaud scans at 60 Hz**

There remains a ton of jitter from the async (maybe readyRead())?, sometimes the viewer-side latency is 20ms.

As a point of reference for what is technically possible, the minimal example [test-device-ping360.cpp](https://github.com/jaxxzer/ping-cpp/blob/test-device/test/test-device-ping360.cpp) has a full scan time of 3.6 seconds (110Hz) and the viewer-side latency is always less than 1ms.