@jaxxzer 2019-08-06
fix #650