@patrickelectric 2019-08-06
@jaxxzer If you can, please break this in multiple PRs to make it easier to master.