@jaxxzer 2019-08-07
fix #657