@jaxxzer 2019-08-07
> @jaxxzer If you can, please break this in multiple PRs to make it easier to master.

I cannot. This patch only works when applied as a whole.

I have updated the commit messages and comments to try to make it easy. Please let me know if you have any trouble understanding.

The hotfix is applied to get around #656