@jaxxzer 2019-08-07
`master` has scan speed of 30Hz

This branch has scan speed of 60Hz.