@jaxxzer 2019-08-07
@patrickelectric Please check also the updated [pr description](https://github.com/bluerobotics/ping-viewer/pull/540#issue-296843485).